Poll fetch
Contorting the RQ API a bit here to essentially poll a mutation. If you squint it does kind of make sense to think of fetch more like a query.
